% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombinePolarities.R
\name{CombinePolarities}
\alias{CombinePolarities}
\title{Combine positive and negative matrices}
\usage{
CombinePolarities(POSITIVE, NEGATIVE, CommonEntitiesFiltered)
}
\arguments{
\item{POSITIVE}{A matrix with positive entities information (Compound Name, Mass, RT, and multiple columns with the area of the compound in samples)}

\item{NEGATIVE}{A matrix with positive entities information (Compound Name, Mass, RT, and multiple columns with the area of the compound in samples)}

\item{CommonEntitiesFiltered}{The data set generated with the FilterbyRT function.}
}
\description{
Take positive and negative matrices and combine them by deleting redundat entities.
}
\examples{
\dontrun{
CombinePolarities(POSITIVE,NEGATIVE,CommonEntitiesFiltered)
}
}

