% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterbyRT.R
\name{FilterbyRT}
\alias{FilterbyRT}
\title{Filter by RT residuals}
\usage{
FilterbyRT(CommonEntitiesImproved, MaxResidual, MinResidual)
}
\arguments{
\item{CommonEntitiesImproved}{Data set resulted from the RemoveMismatch function}

\item{MaxResidual}{Maximum residual allowed for RT+ vs RT- association}

\item{MinResidual}{Minimum residual allowed for RT+ vs RT- association}
}
\value{
Plot filtered (RT+ vs RT-, regression, "residuals vs predicted", and Q-Q plot)

New CommonEntities table filtered, obtained after removing entities with very high or low residuals or RT+ vs RT-.
}
\description{
Remove those entities with residuals above and below a maximum and minimum specified value.
}
\examples{
\dontrun{
CommonEntitiesFiltered<-FilterbyRT(CommonEntitiesImproved,MaxResidual=0.5,MinResidual=(-0.5))
}
}

