% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveMismatch.R
\name{RemoveMismatch}
\alias{RemoveMismatch}
\title{Remove mismatched entities}
\usage{
RemoveMismatch(CommonEntities)
}
\arguments{
\item{CommonEntities}{(Matrix generated from the FindCommon function).}
}
\value{
CommonEntitiesImproved The matrix without mismatched entities.
}
\description{
Remove those entities that have been associated to more than one adduct, retaining only the most probable.
}
\examples{
\dontrun{
CommonEntitiesImproved<-RemoveMismatch(CommonEntities)
}
}

