% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudyRTdiff.R
\name{StudyRTdiff}
\alias{StudyRTdiff}
\title{Study RT differences to lately remove outliers}
\usage{
StudyRTdiff(CommonEntitiesImproved)
}
\arguments{
\item{CommonEntitiesImproved}{The resultant data set from the function RemoveMismatch}
}
\value{
Plot (RT+ vs RT-, regression, "residuals vs predicted", and Q-Q plot)

The CommonEntitiesImproved dataset now included a new column with residuals of each entity for the RT+ vs RT- regression.
}
\description{
Study the correlation between RT in positive and negative ionization modes to find those entities that have been associated wrongly.
}
\examples{
\dontrun{
CommonEntitiesImproved<-StudyRTdiff(CommonEntitiesImproved)
}
}

